-Suppress many of the warnings when building in debug mode with gcc. --This involves fixing signed/unsigned type mismatches, reordering member construction lists, removing "inline" from function definitions, etc...

+++ Source/Ember/DensityFilter.h
@@ -108,7 +108,7 @@ public:
 		size_t w;
 		int intFilterCount, maxIndex;
 		int rowSize;
-		size_t filterLoop;
+        int filterLoop;
 		int keepThresh = 100;
 		uint filterCoefIndex = 0;
 		T decFilterCount;
